+++ app/src/main/java/com/surivalcoding/composerecipeapp/presentation/signup/SignUpScreen.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp.presentation.screen
+package com.surivalcoding.composerecipeapp.presentation.signup
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -30,6 +30,8 @@ import androidx.compose.ui.unit.dp
 import com.surivalcoding.composerecipeapp.presentation.component.BigButton
 import com.surivalcoding.composerecipeapp.presentation.component.InputField
 import com.surivalcoding.composerecipeapp.presentation.component.SocialLogin
+import com.surivalcoding.composerecipeapp.presentation.signin.OrSignInDivider
+import com.surivalcoding.composerecipeapp.presentation.signin.TextWithLinkButton
 import com.surivalcoding.composerecipeapp.ui.theme.AppColors
 import com.surivalcoding.composerecipeapp.ui.theme.AppTextStyles
 import com.surivalcoding.composerecipeapp.ui.theme.ComposeRecipeAppTheme
